Finish Friend/Friend Request process
A user is now able to send a friend request, view all of their own received requests, and accept them, creating friends in the process. Also the root page is now a user's homepage, which is just user#show, which grabs its user from the current_user. I'll probably need to change this, maybe to a new Homepage resource or smth.